add missing package declaration to Pod::Confluence::TestUtil
The test module Pod::Confluence::TestUtil did not have a package
statement, so it didn't work as a proper exporter and instead just
dumped its subs in whatever namespace was active when first using the
module. Convert it to a real package and fix the imports.

Without the package statement, the module was not set up as an exporter
properly, so it had no import method. The arguments to the use statement
were being ignored by perl. Future versions of perl are intending to
make passing arguments to an undefined import method an error.